refactor: unify file names to kebab-case and fix imports
