+++ project-consensus-frontend/src/app/about/page.tsx
@@ -5,7 +5,7 @@ import { SiteNavigation } from '@/components/SiteNavigation';
 import { Card, CardHeader, CardTitle, CardDescription, CardContent, CardFooter } from '@/components/ui/card';
 import { Alert, AlertDescription, AlertTitle } from '@/components/ui/alert';
 import { Button } from '@/components/ui/button';
-import { useI18n } from '@/hooks/useI18n';
+import { useI18n } from '@/hooks/use-i18n';
 import Link from 'next/link';
 
 export default function AboutPage() {
