+++ project-consensus-frontend/src/components/ForumPostCommentCard.tsx
@@ -19,7 +19,7 @@ import {
   DialogTitle,
 } from "@/components/ui/dialog";
 
-import { useI18n } from "@/hooks/useI18n";
+import { useI18n } from "@/hooks/use-i18n";
 import { sanitizeHtml } from "@/lib/html-utils";
 import { cn } from "@/lib/utils";
 import type { ForumPostComment } from "@/types/forum";
@@ -81,7 +81,7 @@ export function ForumPostCommentCard({
     }
     deleteEventHandlersRef.current = {};
   }, []);
-  
+
   // 用于存储回复删除前的原始状态，支持删除操作的撤销功能
   // Stores the original state of replies before deletion to support rollback functionality for delete operations
   const prevRepliesByIdRef = React.useRef<Map<string, ForumPostComment>>(new Map());
@@ -191,11 +191,11 @@ export function ForumPostCommentCard({
     setIsRepliesLoading(true);
     setRepliesError(null);
     try {
-      const data = await fetchForumComments({ 
-        postId: comment.postId, 
-        replyTo: comment.id, 
-        page: 1, 
-        pageSize: 5 
+      const data = await fetchForumComments({
+        postId: comment.postId,
+        replyTo: comment.id,
+        page: 1,
+        pageSize: 5
       });
       setReplies(data.results);
       setRepliesNextUrl(toRelative(data.next));
@@ -216,11 +216,11 @@ export function ForumPostCommentCard({
       // This is a simplified approach - in a real app you might want to store page numbers
       let currentPage = 2; // Start from page 2 since we already loaded page 1
       while (repliesNextUrl) {
-        const data = await fetchForumComments({ 
-          postId: comment.postId, 
-          replyTo: comment.id, 
-          page: currentPage, 
-          pageSize: 5 
+        const data = await fetchForumComments({
+          postId: comment.postId,
+          replyTo: comment.id,
+          page: currentPage,
+          pageSize: 5
         });
         setReplies(prev => {
           const existing = new Set((prev ?? []).map(c => c.id));
@@ -355,9 +355,9 @@ export function ForumPostCommentCard({
         <div className="flex-1 min-w-0 overflow-hidden">
           <div className="flex items-center gap-2 mb-1 flex-wrap">
             <span className="font-medium text-sm text-foreground">
-              {comment.isAnonymous 
-                ? (currentUserId && comment.author.id === currentUserId 
-                    ? `${comment.author.name} (${t('common.anonymous')})` 
+              {comment.isAnonymous
+                ? (currentUserId && comment.author.id === currentUserId
+                    ? `${comment.author.name} (${t('common.anonymous')})`
                     : t('common.anonymous'))
                 : comment.author.name}
               {currentUserId && comment.author.id === currentUserId && (
@@ -376,12 +376,12 @@ export function ForumPostCommentCard({
               title={parentComment.isDeleted ? t('comment.deleted') : stripHtmlTags(parentComment.content)}
             >
               <span className="font-medium flex-shrink-0">
-                {t('comment.repliesTo', { 
-                  name: parentComment.isAnonymous 
-                    ? (currentUserId && parentComment.author.id === currentUserId 
-                        ? `${parentComment.author.name} (${t('common.anonymous')})` 
+                {t('comment.repliesTo', {
+                  name: parentComment.isAnonymous
+                    ? (currentUserId && parentComment.author.id === currentUserId
+                        ? `${parentComment.author.name} (${t('common.anonymous')})`
                         : t('common.anonymous'))
-                    : parentComment.author.name 
+                    : parentComment.author.name
                 })}
                 {currentUserId && parentComment.author.id === currentUserId && (
                   <span className="text-muted-foreground"> ({t('common.me')})</span>
@@ -539,9 +539,9 @@ export function ForumPostCommentCard({
                           title={r.isDeleted ? t('comment.deleted') : stripHtmlTags(r.content)}
                         >
                           <span className="font-medium flex-shrink-0">
-                            {r.isAnonymous 
-                              ? (currentUserId && r.author.id === currentUserId 
-                                  ? `${r.author.name} (${t('common.anonymous')})` 
+                            {r.isAnonymous
+                              ? (currentUserId && r.author.id === currentUserId
+                                  ? `${r.author.name} (${t('common.anonymous')})`
                                   : t('common.anonymous'))
                               : r.author.name}
                             {currentUserId && r.author.id === currentUserId && (
