+++ project-consensus-frontend/src/components/SearchBar.tsx
@@ -2,7 +2,7 @@ import React, { useState, FormEvent, useEffect } from 'react';
 import { Search, X } from 'lucide-react';
 import { cn } from '@/lib/utils';
 import { Input } from '@/components/ui/input';
-import { useDebounce } from '@/hooks/useDebounce';
+import { useDebounce } from '@/hooks/use-debounce';
 
 interface SearchBarProps {
   placeholder?: string;
