+++ project-consensus-frontend/src/data/sample-courses.ts
