+++ project-consensus-frontend/src/data/sample-curriculum.ts
